gemini/requirements.txt
flask
flask-cors
python-dotenv
google-genai
fer
opencv-python
tensorflow
numpy
